New article on forensic phonetics.
Title: "The long-term average spectrum in forensic phonetics: From collation to discrimination of speakers"

Abstract: "The present work aims to provide more data regarding the comparison and discrimination of speech samples in the field of forensic phonetics. The methodology used was an acoustic method called long-term average spectrum (LTAS), by calculating its spectral decline. The informants were grouped into a total of four female speakers of the same age by close geographical origins. The extracted data were presented qualitatively and quantitatively. In view of the results, the identification of speakers was not 
successful in all cases. After discussing the results, implications for judicial phonetics and recommendations for future studies were collected."